Remove integral convergence tests from unit tests
