+++ test/Spaces/distributed_cuda/ddss4.jl
@@ -100,7 +100,7 @@ pid, nprocs = ClimaComms.init(context)
     end
     p = @allocated Spaces.weighted_dss!(y0, dss_buffer)
     if pid == 1
-        @test p ≤ 6752
+        @test p ≤ 7008
     end
 
 end
